@@ -1,10 +1,10 @@
-using System;
-using System.Threading.Tasks;
-using Orleans;
 using SignalR.Orleans.Core;
 
 namespace SignalR.Orleans.Clients
 {
+    /// <summary>
+    /// A single connection
+    /// </summary>
     public interface IClientGrain : IHubMessageInvoker, IGrainWithStringKey
     {
         Task OnConnect(Guid serverId);
