@@ -1,22 +1,28 @@
 using System;
+using Orleans.Streams;
 
 namespace SignalR.Orleans
 {
     public static class Constants
     {
-        public const string PUBSUB_PROVIDER = "PubSubStore";
-        // todo: ideally it doesnt use the default name so consumers can replace the provider and not affecting the default - it will be breaking tho.
-        //public const string PUBSUB_PROVIDER = "ORLEANS_SIGNALR_PUBSUB_PROVIDER";
+        /// <summary>
+        /// Name of the streams storage provider used by the signalR orleans backplane grain streams.
+        /// </summary>
+        public const string PUBSUB_PROVIDER = "ORLEANS_SIGNALR_PUBSUB_PROVIDER";
 
+        /// <summary>
+        /// Name of the state storage provider used by signalR orleans backplane grains.
+        /// </summary>
         public const string STORAGE_PROVIDER = "ORLEANS_SIGNALR_STORAGE_PROVIDER";
 
-        public const string SERVERS_STREAM = "SERVERS_STREAM";
-        public const string SERVER_DISCONNECTED = "SERVER_DISCONNECTED";
+        /// <summary>
+        /// Name used to access the <see cref="IStreamProvider"/> that supplies signalR orleans backplane streams.
+        /// </summary>
         public const string STREAM_PROVIDER = "ORLEANS_SIGNALR_STREAM_PROVIDER";
-        public static readonly Guid CLIENT_DISCONNECT_STREAM_ID = Guid.Parse("bdcff7e7-3734-48ab-8599-17d915011b85");
-        public static readonly Guid ALL_STREAM_ID = Guid.Parse("fbe53ecd-d896-4916-8281-5571d6733566");
 
-        internal const int HEARTBEAT_PULSE_IN_MINUTES = 30;
-        internal const int SERVERDIRECTORY_CLEANUP_IN_MINUTES = HEARTBEAT_PULSE_IN_MINUTES * 3;
+        /// <summary>
+        /// The number of minutes that each signalR hub server must pulse the server directory.
+        /// </summary>
+        internal const int SERVER_HEARTBEAT_PULSE_IN_MINUTES = 30;
     }
 }
